Feature/initcode size limit using opcodes contract creation (#2697)
* Initcode validation for opcodes contract creation

- Adding scenarios of test for the contract creation with DSL
- The validation for the initCode size test was added
- The validation is based on the RSKIP that enables it
- Validation added for the contractCreation during transaction execution
- In order to have the exact same logic for the initcode cost calculation,
we introduced some new classes that can be easily reused by the TransactionExecutor and VM classes. This way, if this logic change,
we won't have to change the code in multiple places.

It was also introduced an interface for it, so we can expand this idea and
in the future, be able to calculate total transactions just calling a method
to calculate it independent of the logic used.

- Adding scenarios of test for the contract creation with DSL
- The validation is based on the RSKIP that enables it

* Added the remaining test scenarios for contracts with opcode creation

* Applying suggestions from review

* Applying some more suggestions from review
---------

+++ rskj-core/src/main/java/org/ethereum/cost/CostCalculator.java
@@ -23,5 +23,5 @@ import org.ethereum.config.blockchain.upgrades.ActivationConfig;
 
 public interface CostCalculator {
 
-    long calculateCost(int dataLength, ActivationConfig.ForBlock activations);
+    long calculateCost(long dataLength, ActivationConfig.ForBlock activations);
 }
