+++ rskj-core/src/main/java/org/ethereum/cost/InitcodeCostCalculator.java
@@ -37,7 +37,7 @@ public class InitcodeCostCalculator implements CostCalculator {
     }
 
     @Override
-    public long calculateCost(int dataLength, ActivationConfig.ForBlock activations) {
+    public long calculateCost(long dataLength, ActivationConfig.ForBlock activations) {
         if ( dataLength > 0 && activations.isActive(ConsensusRule.RSKIP438) ) {
             return  INITCODE_WORD_COST  *  ( (long) Math.ceil( ( (double) dataLength ) / 32 ) );
         }
